ml_server/requirements.txt
# Orafinite ML Sidecar Dependencies

# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2

# LLM Guard - Real-time protection
llm-guard==0.3.14

# Garak - Vulnerability scanning
garak==0.9.0.13

# NeMo Guardrails (optional, for advanced dialog control)
# nemoguardrails==0.10.0

# ML/AI
torch>=2.0.0
transformers>=4.36.0
numpy>=1.24.0
optimum[onnxruntime]>=1.16.0

# Async support
asyncio
aiofiles

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
